fix: fix timestamp parsing value in CrashInfo (Android)

+++ packages/trip-analysis/android/src/main/java/com/reactnativedrivekittripanalysis/TripListenerMappers.kt
@@ -65,7 +65,7 @@ fun mapCrashStatus(status: CrashStatus): String {
 fun mapDKCrashInfo(info: DKCrashInfo): ReadableMap {
   val rnCrashInfo = Arguments.createMap()
   rnCrashInfo.putString("crashId", info.crashId)
-  rnCrashInfo.putInt("timestamp", info.date.time.toInt())
+  rnCrashInfo.putDouble("timestamp", info.date.time.toDouble())
   rnCrashInfo.putInt("probability", info.probability.toInt())
   rnCrashInfo.putDouble("latitude", info.latitude)
   rnCrashInfo.putDouble("longitude", info.longitude)
